Resolved test coverage issue.
I removed the instrumentation/coverage-generation flags from
Template Framework Project.

Then I setup the Podfile for the tests to respect a custom
"Coverage" build configuration, based on debug.

Then I added "Coverage" as a custom build configuration
(duplicated from "Debug") for both TestsProject and the
framework project settings. For these "Coverage" build
configurations, I set the instrumentation/coverage-generation
flags to YES and for the Release and Debug configurations I set
them to NO.

Finally, I made "Coverage" be the build configuration used for
the "Test" action on the TestsProject and framework schemes,
as well as the "Run" action on the TestsProject scheme.

This seems to have achieved the Template Framework Project
building okay in Debug and Release configurations and the
test coverage information still being generated during the
running of our unit tests on TestsProject.